Seems to be working version
-Added a cell at the end of y to detect event when particles nucleate to
empty section. In that case, the diameter of the section is changed to
correspond the diameter of nucleating particles.
-If a section grows over limit and is emptied, its diameter is reset to
the original value. If this wouldn't be done and particles are added in
this section by coagulation, the diameter grows immediatly over the
limit by condensation. This causes too many events and ode does not work
always.
-Some results comparing the old simulations to the recent nucleation
mechanism.